+++ src/app/clients/clients-view/client-actions/reactivate-client/reactivate-client.component.ts
@@ -14,10 +14,9 @@ import { SettingsService } from 'app/settings/settings.service';
 @Component({
   selector: 'mifosx-reactivate-client',
   templateUrl: './reactivate-client.component.html',
-  styleUrls: ['./reactivate-client.component.scss']
+  styleUrls: ['./reactivate-client.component.scss'],
 })
 export class ReactivateClientComponent implements OnInit {
-
   /** Minimum date allowed. */
   minDate = new Date(2000, 0, 1);
   /** Maximum date allowed. */
@@ -27,6 +26,8 @@ export class ReactivateClientComponent implements OnInit {
   /** Client Account Id */
   clientId: any;
 
+  reactivateData: any;
+
   /**
    * @param {FormBuilder} formBuilder Form Builder
    * @param {clientsService} clientsService Clients Service
@@ -35,12 +36,17 @@ export class ReactivateClientComponent implements OnInit {
    * @param {Router} router Router
    * @param {SettingsService} settingsService Setting service
    */
-  constructor(private formBuilder: UntypedFormBuilder,
-              private clientsService: ClientsService,
-              private dateUtils: Dates,
-              private route: ActivatedRoute,
-              private router: Router,
-              private settingsService: SettingsService) {
+  constructor(
+    private formBuilder: UntypedFormBuilder,
+    private clientsService: ClientsService,
+    private dateUtils: Dates,
+    private route: ActivatedRoute,
+    private router: Router,
+    private settingsService: SettingsService
+  ) {
+    this.route.data.subscribe((data: { clientActionData: any }) => {
+      this.reactivateData = data.clientActionData.narrations;
+    });
     this.clientId = this.route.parent.snapshot.params['clientId'];
   }
 
@@ -56,7 +62,8 @@ export class ReactivateClientComponent implements OnInit {
    */
   createReactivateClientForm() {
     this.reactivateClientForm = this.formBuilder.group({
-      'reactivationDate': ['', Validators.required]
+      reactivationDate: ['', Validators.required],
+      reactivationReasonId: ['', Validators.required],
     });
   }
 
@@ -69,17 +76,16 @@ export class ReactivateClientComponent implements OnInit {
     const locale = this.settingsService.language.code;
     const dateFormat = this.settingsService.dateFormat;
     const prevReactivationDate: Date = this.reactivateClientForm.value.reactivationDate;
-    if (reactivateClientFormData.closureDate instanceof Date) {
+    if (reactivateClientFormData.reactivationDate instanceof Date) {
       reactivateClientFormData.reactivationDate = this.dateUtils.formatDate(prevReactivationDate, dateFormat);
     }
     const data = {
       ...reactivateClientFormData,
       dateFormat,
-      locale
+      locale,
     };
     this.clientsService.executeClientCommand(this.clientId, 'reactivate', data).subscribe(() => {
       this.router.navigate(['../../'], { relativeTo: this.route });
     });
   }
-
 }