+++ src/app/core/http/api-prefix.interceptor.ts
@@ -7,6 +7,7 @@ import { Observable } from 'rxjs';
 
 /** Environment Configuration */
 import { SettingsService } from 'app/settings/settings.service';
+import { HttpCacheService } from './http-cache.service';
 
 /**
  * Http request interceptor to prefix a request with `serverUrl`.
@@ -16,7 +17,7 @@ export class ApiPrefixInterceptor implements HttpInterceptor {
   /**
    * @param {SettingsService} settingsService Settings Service
    */
-   constructor(private settingsService: SettingsService) {}
+  constructor(private settingsService: SettingsService, private cacheService: HttpCacheService) {}
 
   /**
    * Intercepts a Http request and prefixes it with `serverUrl`.
@@ -25,11 +26,14 @@ export class ApiPrefixInterceptor implements HttpInterceptor {
     /**
      * Ignore URLs that are complete for i18n
      */
-     if (!request.url.includes('http:') && !request.url.includes('https:')) {
+    if (!request.url.includes('http:') && !request.url.includes('https:')) {
       request = request.clone({ url: this.settingsService.serverUrl + request.url });
+    }
+
+    if (request.method === 'PUT' || request.method === 'POST' || request.method === 'DELETE') {
+      this.cacheService.cleanCache();
     }
 
     return next.handle(request);
   }
-
 }